Ajout de swagger et prise en comptes des logs par un simple AOP et les testes mocké avec mockito

+++ src/main/java/com/fr/register/service/UserServiceImpl.java
@@ -4,7 +4,6 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import com.fr.register.dto.UserDto;
@@ -14,55 +13,78 @@ import com.fr.register.model.UserModel;
 import com.fr.register.repository.UserRepository;
 
 @Service
-public class UserServiceImpl implements UserServiceI{
+public class UserServiceImpl implements UserServiceI {
+
+	private UserRepository userRepository;
 
 	@Autowired
-	UserRepository userRepository ;
-	
-	@Autowired
-	UserDto dto ;
-	
+	public void setUserRepository(UserRepository userRepository) {
+		this.userRepository = userRepository;
+	}
+
+	UserDto dto = new UserDto();
+
 	@Override
 	public List<UserModel> getAllUser() {
-		return dto.toModels(userRepository.findAll());
+		List<User> userList = userRepository.findAll();
+
+		return dto.toModels(userList);
 	}
 
 	@Override
 	public UserModel getUserById(Long id) {
 		Optional<User> optionalUser = userRepository.findById(id);
-		if(optionalUser.isPresent()) {
+		if (optionalUser.isPresent()) {
 			return dto.toModel(optionalUser.get());
 
-		}else {
-			return null ;
+		} else {
+			return null;
 		}
 	}
 
 	@Override
 	public UserModel saveUser(UserModel model) {
-		return dto.toModel(userRepository.save(dto.toEntity(model)));
+		User u = null;
+		try {
+			u = userRepository.save(dto.toEntity(model));
+
+		} catch (Exception e) {
+			e.printStackTrace();
+		}
+		return dto.toModel(u);
 	}
 
 	@Override
-	public void deleteUserById(Long id) throws UserNotFoundException {
-		try {
+	public UserModel updateUserById(Long id, UserModel user) throws UserNotFoundException {
+		userRepository.findById(id).orElseThrow(() -> new UserNotFoundException(id));
+		user.setId(id);
+		return dto.toModel(userRepository.save(dto.toEntity(user)));
+	}
+
+	@Override
+	public UserModel deleteUserById(Long id) throws UserNotFoundException {
+
+		UserModel user = null;
+		user = getUserById(id);
+		if (null != user) {
 			userRepository.deleteById(id);
-		} catch (EmptyResultDataAccessException ex) {
-			throw new UserNotFoundException();
+			return user;
+		} else {
+			return user;
 		}
-		
+
 	}
 
 	@Override
 	public UserModel getUserByName(String name) {
 		Optional<User> optionalUser = userRepository.findByName(name);
-		if(optionalUser.isPresent()) {
+		if (optionalUser.isPresent()) {
 			return dto.toModel(optionalUser.get());
 
-		}else {
-			return null ;
+		} else {
+			return null;
 		}
-		
+
 	}
 
 }
